[test] 기능 테스트 추가

+++ src/main/java/lotto/domain/LottoResult.java
@@ -65,7 +65,7 @@ public class LottoResult {
             return NO_BENEFIT.getSetting();
         }
 
-        return (totalReward / totalSpendAmount) * PERCENTAGE_INDICATOR.getSetting();
+        return (totalReward / (double) totalSpendAmount) * PERCENTAGE_INDICATOR.getSetting();
     }
